LayoutTests/media/video-canvas-createPattern-expected.txt


EVENT(canplaythrough)
EVENT(seeked)
Saw close enough to (200, 200, 0) at (40, 165) OK
Saw close enough to (200, 200, 0) at (355, 165) OK
Saw close enough to (200, 200, 0) at (40, 403) OK
Saw close enough to (200, 200, 0) at (355, 403) OK
Saw close enough to (57, 202, 203) at (56, 165) OK
Saw close enough to (57, 202, 203) at (371, 165) OK
Saw close enough to (57, 202, 203) at (56, 403) OK
Saw close enough to (57, 202, 203) at (371, 403) OK
Saw close enough to (53, 203, 9) at (73, 165) OK
Saw close enough to (53, 203, 9) at (388, 165) OK
Saw close enough to (53, 203, 9) at (73, 403) OK
Saw close enough to (53, 203, 9) at (388, 403) OK
EVENT(seeked)
Saw close enough to (200, 200, 0) at (40, 165) OK
Saw close enough to (200, 200, 0) at (355, 165) OK
Saw close enough to (200, 200, 0) at (40, 403) OK
Saw close enough to (200, 200, 0) at (355, 403) OK
Saw close enough to (57, 202, 203) at (56, 165) OK
Saw close enough to (57, 202, 203) at (371, 165) OK
Saw close enough to (57, 202, 203) at (56, 403) OK
Saw close enough to (57, 202, 203) at (371, 403) OK
Saw close enough to (53, 203, 9) at (73, 165) OK
Saw close enough to (53, 203, 9) at (388, 165) OK
Saw close enough to (53, 203, 9) at (73, 403) OK
Saw close enough to (53, 203, 9) at (388, 403) OK
EVENT(seeked)
Saw close enough to (200, 200, 0) at (40, 165) OK
Saw close enough to (200, 200, 0) at (355, 165) OK
Saw close enough to (200, 200, 0) at (40, 403) OK
Saw close enough to (200, 200, 0) at (355, 403) OK
Saw close enough to (57, 202, 203) at (56, 165) OK
Saw close enough to (57, 202, 203) at (371, 165) OK
Saw close enough to (57, 202, 203) at (56, 403) OK
Saw close enough to (57, 202, 203) at (371, 403) OK
Saw close enough to (53, 203, 9) at (73, 165) OK
Saw close enough to (53, 203, 9) at (388, 165) OK
Saw close enough to (53, 203, 9) at (73, 403) OK
Saw close enough to (53, 203, 9) at (388, 403) OK
EVENT(seeked)
Saw close enough to (200, 200, 0) at (40, 165) OK
Saw close enough to (200, 200, 0) at (355, 165) OK
Saw close enough to (200, 200, 0) at (40, 403) OK
Saw close enough to (200, 200, 0) at (355, 403) OK
Saw close enough to (57, 202, 203) at (56, 165) OK
Saw close enough to (57, 202, 203) at (371, 165) OK
Saw close enough to (57, 202, 203) at (56, 403) OK
Saw close enough to (57, 202, 203) at (371, 403) OK
Saw close enough to (53, 203, 9) at (73, 165) OK
Saw close enough to (53, 203, 9) at (388, 165) OK
Saw close enough to (53, 203, 9) at (73, 403) OK
Saw close enough to (53, 203, 9) at (388, 403) OK
END OF TEST
